+++ src/filters/ttml_conv.c
@@ -42,7 +42,7 @@ typedef struct
 
 	u8 *buf;
 	u32 buf_alloc;
-	s64 last_end_ts;
+	u64 last_end_ts;
 
 	Bool srt_conv;
 } TTMLConvCtx;
@@ -255,7 +255,7 @@ GF_Err ttmlconv_process(GF_Filter *filter)
 				goto exit;
 			}
 			ttmlconv_dump_node(ctx, p, dump);
-			u32 size = gf_ftell(dump);
+			u32 size = (u32) gf_ftell(dump);
 			gf_fseek(dump, 0, SEEK_SET);
 			u8 *output;
 			dst_pck = gf_filter_pck_new_alloc(ctx->opid, size, &output);
@@ -271,7 +271,7 @@ GF_Err ttmlconv_process(GF_Filter *filter)
 			gf_filter_pck_set_sap(dst_pck, GF_FILTER_SAP_1);
 			gf_filter_pck_set_byte_offset(dst_pck, GF_FILTER_NO_BO);
 			gf_filter_pck_set_cts(dst_pck, p_start_ts);
-			gf_filter_pck_set_duration(dst_pck, p_end_ts-p_start_ts);
+			gf_filter_pck_set_duration(dst_pck, (u32) (p_end_ts-p_start_ts));
 			gf_filter_pck_send(dst_pck);
 		}
 	}
